+++ src/main/kotlin/com/JuMarket/autoatendimento/JuMarket/repository/CartRepository.kt
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository
 
 
 @Repository
-interface CartRepository : JpaRepository<Cart, Long>
+interface CartRepository : JpaRepository<Cart, Int>
