+++ src/main/kotlin/com/JuMarket/autoatendimento/JuMarket/repository/SaleRepository.kt
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface SaleRepository : JpaRepository<Sale, Long>
+interface SaleRepository : JpaRepository<Sale, Int>
