+++ src/main/kotlin/com/JuMarket/autoatendimento/JuMarket/service/IProductService.kt
@@ -4,8 +4,8 @@ import com.JuMarket.autoatendimento.JuMarket.entity.Product
 
 interface IProductService {
     fun save(product: Product): Product
-    fun delete(id: Long)
-    fun findById(id: Long): Product
+    fun delete(id: Int)
+    fun findById(id: Int): Product
 
     fun findAll(): List<Product>
 }
